[BDN-3] Create dummy auction endpoint
