+++ bidon_api/app/controllers/application_controller.rb
@@ -1,6 +1,12 @@
 class ApplicationController < ActionController::API
   before_action :validate_bidon_header!
 
+  rescue_from StandardError do |_e|
+    render json: { error: { code: 500, message: 'Internal Server Error' } }, status: :internal_server_error
+  end
+
+  private
+
   def validate_bidon_header!
     return if request.env['X-BidOn-Version'].present?
 
@@ -8,7 +14,11 @@ class ApplicationController < ActionController::API
            status: :unprocessable_entity
   end
 
-  rescue_from StandardError do |_e|
-    render json: { error: { code: 500, message: 'Internal Server Error' } }, status: :internal_server_error
+  def render_empty_result
+    render json: { success: true }, status: :ok
+  end
+
+  def render_app_key_invalid
+    render json: { error: { code: 422, message: 'App key is invalid' } }, status: :unprocessable_entity
   end
 end
