+++ bidon_api/config/routes.rb
@@ -1,3 +1,4 @@
 Rails.application.routes.draw do
   post 'config', to: 'config#create'
+  post 'auction', to: 'auction#create'
 end
